Shortest path grapht
Builds CFG of all program locations and computes the shortest distance from all locations in the graph to a single property.
WARNING: if more than one property is given, we use the first one only.

Writes these distances to the corresponding node in a locs structure passed as argument to the constructor, in order for this to be used to guide the symex search